backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
pymysql
cryptography
python-dotenv
passlib[bcrypt]
PyJWT
email-validator
